refactor: 커뮤니티 게시글 폼을 공통 UI와 생성 로직으로 분리 #1492
