+++ frontend/src/common/types/communityPostForm.ts
@@ -1,0 +1,7 @@
+export interface CommunityPostFormValues {
+  title: string;
+  content: string;
+  isAnonymous: boolean;
+  nickname?: string;
+  guestPassword?: string;
+}
